+++ src/main/java/com/se/apiserver/v1/post/application/service/PostUpdateService.java
@@ -12,8 +12,8 @@ import com.se.apiserver.v1.common.domain.exception.BusinessException;
 import com.se.apiserver.v1.post.domain.entity.Post;
 import com.se.apiserver.v1.post.domain.entity.PostTagMapping;
 import com.se.apiserver.v1.post.application.error.PostErrorCode;
-import com.se.apiserver.v1.post.infra.dto.PostCreateDto;
-import com.se.apiserver.v1.post.infra.dto.PostUpdateDto;
+import com.se.apiserver.v1.post.application.dto.PostCreateDto;
+import com.se.apiserver.v1.post.application.dto.PostUpdateDto;
 import com.se.apiserver.v1.attach.infra.repository.AttachJpaRepository;
 import com.se.apiserver.v1.post.infra.repository.PostJpaRepository;
 import com.se.apiserver.v1.tag.application.error.TagErrorCode;
