+++ src/main/java/com/se/apiserver/v1/post/infra/api/PostApiController.java
@@ -6,9 +6,9 @@ import com.se.apiserver.v1.post.application.service.PostCreateService;
 import com.se.apiserver.v1.post.application.service.PostDeleteService;
 import com.se.apiserver.v1.post.application.service.PostReadService;
 import com.se.apiserver.v1.post.application.service.PostUpdateService;
-import com.se.apiserver.v1.post.infra.dto.PostCreateDto;
-import com.se.apiserver.v1.post.infra.dto.PostReadDto;
-import com.se.apiserver.v1.post.infra.dto.PostUpdateDto;
+import com.se.apiserver.v1.post.application.dto.PostCreateDto;
+import com.se.apiserver.v1.post.application.dto.PostReadDto;
+import com.se.apiserver.v1.post.application.dto.PostUpdateDto;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
